Update EventsToday to use separate EventLists
